+++ tests/test_events.py
@@ -3,11 +3,10 @@ from __future__ import annotations
 import pytest
 
 from pyflp._events import AsciiEvent, EventEnum, EventTree, U8Event, WORD
-from pyflp.exceptions import EventIDOutOfRange, InvalidEventChunkSize
 
 
 def test_id_out_of_range():
-    with pytest.raises(EventIDOutOfRange, match=str(tuple(range(0, WORD)))):
+    with pytest.raises(ValueError, match=str(tuple(range(0, WORD)))):
         U8Event(EventEnum(128), b"\x00")
 
     with pytest.raises(ValueError):
@@ -15,7 +14,7 @@ def test_id_out_of_range():
 
 
 def test_invalid_chunk_size():
-    with pytest.raises(InvalidEventChunkSize, match="1"):
+    with pytest.raises(BufferError, match="1"):
         U8Event(EventEnum(0), b"12")